Use python to code the Cooley–Tukey Fast Fourier transform algorithm from first principles.
The Cooley Tucker Fast Forrier Transform
The Fourier transform converts a function from time domain to frequency domain. This transformation is commonly used for audio and image processing but it is also used in various other applications. However, most of these applications work with the information in a discrete format. The discrete Fourier transform[1] (DFT) is given with the following equation:




Here, k ranges from 0 to N-1. This formula is inefficient to compute, especially when considering real-time applications of this algorithm. In order to address this problem, the Cooley–Tukey Fast Fourier transform algorithm incorporates a few mathematical “tricks”. The first step in simplifying this equation is to split the formula into even and odd values of n as shown below: